use outermost IP DSCP for stream classification
Using the outermost IP DSCP for stream classification makes ROHC streams
more robust to QoS reordering: all packets of one ROHC stream are
transmitted with the same priority, so they are not reordered.

Using the outermost IP DSCP for stream classification may decrease
compression efficiency a little bit: a variation of DSCP causes a
new stream to be created.