Q&A: RFC3095: remove dead code related to send_static
The only STATIC field for the RFC3095-based profiles is the IPv4
Protocol or IPv6 Next Header field. A change of value requires an IR,
IR-DYN or UOR-2 packet.

The Protocol/NH field is however part of the definition of one ROHC
context, so a change of value would cause a change of context. A
given context can never see a change of Protocol or NH.

The code that checks for Protocol/NH changes and forces an IR, IR-DYN
or UOR-2 packet is therefore never used. This is dead code. It can be
removed safely.

Thanks to yiyuechan for discovering the problem in:
  https://answers.launchpad.net/rohc/+question/675656