Q&A: RFC3095: remove dead code related to RTP SSRC changes
The only STATIC field for the RFC3095 RTP profile is the SSRC field.
A change of value requires an IR or IR-DYN packet.

The RTP SSRC field is however part of the definition of one ROHC RTP
context, so a change of value would cause a change of context. A given
context can never see a change of RTP SSRC.

The code that checks for RTP SSRC changes and forces an IR or IR-DYN
packet is therefore never used. This is dead code. It can be removed
safely.

Thanks to yiyuechan for discovering the problem in:
  https://answers.launchpad.net/rohc/+question/675656