efficiency: TCP: do not transmit TS option in full when not required
The seq_8, rnd_8 and co_common packets may transmit both a compressed
list of TCP options (list structure + optional items) and the irregular
parts of the TCP options.

If the structure of the list changed with the current packet or in the
last few packets, the structure of the list shall be transmitted. One
seq_8, rnd_8 or co_common packet is therefore chosen.

If the TS option changed too, it shall be transmitted too. However, if
the changes are small (<= 29 bits of TS shall be transmitted), the TS
option may be transmitted in the irregular chain. TS encoding in the
irregular chain requires 8, 16, 24 or 32 bits. TS encoding in the base
header of seq_8, rnd_8 or co_common packets always requires 32 bits.

So, if list structure changed and if TS changed not too much, the
compressor should prefer encoding the TS option in the irregular chain
instead of the base header of seq_8, rnd_8 or co_common packets. This
may save up to 3 bytes.