Q&A: remove unused first_used and latest_used timestamps
The first_used timestamps on compressor and decompressor sides are
correctly initialized. They are however never used.

The latest_used timestamp on decompressor side is initialized at context
creation, but not updated when a packet is decompressed. It is not used.

Only the latest_used timestamp on compressor side is correctly
initialized and updated. It is used to determine which context to
recycle if no free context is available (LRU, Least Recently Used
algorithm).

So, let's remove the unused timestamps. It decreases the size of the
context structures in memory. It should help with processing performance.

The problem about latest_used timestamp on decompressor side was
reported by Yakir Matusovsky <yakir.matusovsky@mimomax.com>